fix: refactor asset assignment dialogs and alerts for employee dashboard
